Add files via upload
This EDA case study aims to: 
Identify patterns which indicate if a client has difficulty paying their installments.
Understand the driving factors (or driver variables) behind loan default, i.e. the variables which are strong indicators of default.  
Understand the patterns present in the data pertaining to the loan repaying capabilities of the applicants & how consumer attributes and loan attributes influence the tendency of default.

EDA Analysis would help in:
Denying the loan, reducing the amount of loan, lending (to risky applicants) at a higher interest rate, etc.
Ensuring that the consumers capable of repaying the loan are not rejected. 
The company can utilize this knowledge for its portfolio and risk assessment.